pyproject.toml
[tool.poetry]
name = "verzekeringen.api"
version = "0.1.0"
description = "The api for scouts verzekeringen"
authors = ["Jonas Elst <jonas.elst@inuits.eu>", "Stijn Verholen <boro@inuits.eu>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.2"
mozilla-django-oidc = "^2.0.0" # @TODO
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
django-filter = "^2.4.0"
environs = "^9.3.2"
drf-yasg2 = "^1.19.4"
PyYAML = "^5.4.1"
packaging = "^20.9"
dj-email-url = "^1.0.2"
django-anymail = {extras = ["sendinblue"], version = "^8.4"}
jsonfield = "^3.1.0"
psycopg2-binary = "^2.9.1"
django-storages = "^1.11.1"
boto3 = "^1.18.9"
pdfrw = "^0.4"
# @TODO
#scouts-auth = {git = "ssh://git@gitlab.inuits.io:2224/boro/scouts_auth.git", rev = "master"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py37']
exclude = '''
/(
  | migrations
)/

'''
